refactor(launcher): extract MorePageMixin; main_app 4018 -> 3854 lines
Continues the monolith split (collaborator confirmed done in this area):
More page (_page_more + open_bootstrap / _force_kill_engine /
_open_perf_reports / _collect_diagnostics) moved to
launcher/pages/more_page.py. Dead imports (PageHeader, APP_VERSION) dropped
from main_app. Import/MRO verified with python3.12, full suite green.
